Updated blob update check

+++ .azure/pipelines/scripts/ConvertTo-ManagedImage.ps1
@@ -180,19 +180,6 @@ try {
     Write-Host "Attempting to clean up resources..."
 } finally {
     Write-Host "Cleaning up and removing the managed disk"
-
-    Write-Host "provisioningState=($diskCreateResult.provisioningState)"
-    Write-Host "Clean=$Clean"
-    if (($diskCreateResult.provisioningState) -eq "Succeeded") {
-        Write-Host "Succeeded"
-    }
-    if ($Clean) {
-        Write-Host "Clean"
-    }
-
-
-
-
     if (($diskCreateResult.provisioningState) -eq "Succeeded" -and $Clean) {
         az disk delete --ids $diskCreateResult.id --yes
     } else {
@@ -206,8 +193,17 @@ try {
 
         # Cleaning up blobs associated with working vm images. These are throw away and only needed while Packer is building an image
         $modifiedDate = (Get-Date -AsUTC -Date ((Get-Date).AddDays(-$CleanBlobsDaysOld)) -Format s) + "Z"
-        Write-Host "Cleaning up all blobs in the 'images' container since $modifiedDate"
-        az storage blob delete-batch --source images --auth-mode login --pattern *.vhd --account-name $StorageAccountName --if-unmodified-since $modifiedDate
+        Write-Host "Cleaning up all blobs in the 'images' container since '$modifiedDate'" only if they exist
+
+        [array]$dateResults = $(az storage blob list --container-name images --auth-mode login --account-name $StorageAccountName --query "[?properties.lastModified<``$modifiedDate``].properties.lastModified"  -o tsv)
+        Write-Host "Looking up vhd blobs to delete based on last modified date '$modifiedDate': $dateResults"
+        if ($dateResults.Length -gt 0) {
+            Write-Host "Attempting to delete vhd blobs..."
+            az storage blob delete-batch --source images --auth-mode login --pattern *.vhd --account-name $StorageAccountName --if-unmodified-since $modifiedDate
+        } else {
+            Write-Host "There are no blobs to delete due to not modified since '$modifiedDate'"
+        }
+        
     } else {
         Write-Host "##vso[task.logissue type=warning]Could not clean up the working vhd blobs in the 'images' container. This could be due errors in access or the 'Clean' switch wasn't set."
     }
